reverted previosly commited bug in gameboy timer targetgbprofiling doesn't print anymore opcode numbers

+++ pypy/lang/gameboy/timer.py
@@ -156,5 +156,5 @@ class Clock(object):
         pass
     
     def get_time(self):
-        return (time.time()*1000)
+        return int(time.time()*1000)
